Occupancies are now stored in the Spectral class
Occupancies are used at various points of the module, typically for getting the valence and conduction bands.
Rather than doing this each time, it is easier to just set the occupancies once (less chance of making a mistake as well!)

In a NCM calculation, the bands are not doubly occupied despite nspins being 1 as there are two spinor components.
Unfortunately, there is no way to distinguish a NCM calculation from a spin degenerate one solely from the .bands file
so need to set a flag to ensure occupancies are set correctly.